admin/requirements.txt
boto3==1.26.13
PyYAML==6.0